# prog: optimize foreachArgImpl()