Per profiling, (*GroupArg).Size() calls from this function were one of the hottest paths during the BenchmarkMutate() benchmark.

Some of those calls are made only to issue a runtime panic, which we arguably don't need unless we're testing the code.

After the changes:
```
          │ /tmp/original │              /tmp/new               │
          │    sec/op     │   sec/op     vs base                │
Mutate-36     221.8µ ± 4%   179.0µ ± 3%  -19.31% (p=0.000 n=15)
```